# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Arvan SMS API
  version: 1.0.0
  description: |
    Modular monolith SMS API.

    **Response envelope**
    - Success: `{ "data": ... }`
    - Error: `{ "error": { "code", "message", "fields?" } }`

    **Message costs**
    - OTP: 50
    - Text normal: 20
    - Text express: 30
servers:
  - url: http://localhost:8080
    description: local

tags:
  - name: Health
  - name: Users
  - name: Messages

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      operationId: healthCheck
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by id
      operationId: getUser
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
              example:
                data:
                  id: 1
                  name: Soheil
                  balance: 10000
                  createdAt: "2026-09-16T12:00:00Z"
                  updatedAt: "2026-09-16T12:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /users/{id}/add-balance:
    post:
      tags: [Users]
      summary: Add balance to user
      operationId: addBalance
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBalanceRequest"
            examples:
              ok:
                value: { amount: 5000 }
              invalid:
                value: { amount: 0 }
      responses:
        "200":
          description: Balance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"

  /messages/send/otp:
    post:
      tags: [Messages]
      summary: Send OTP messages
      operationId: sendOTP
      description: |
        Idempotent by `(userId, idempotencyKey)`.
        OTP is converted to text/normal before queueing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOTPRequest"
            example:
              userId: "1"
              idempotencyKey: "otp-123"
              type: express
              template: "Hi {{name}}, code {{code}}"
              recipients:
                - mobile: "09121234567"
                  variables: { name: "Soheil", code: "123456" }
                - mobile: "09351234567"
                  variables: { name: "Ali", code: "654321" }
      responses:
        "200":
          description: Accepted (may include skipped/rejected per-recipient items)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessagesEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /messages/send/text:
    post:
      tags: [Messages]
      summary: Send text messages
      operationId: sendText
      description: |
        Idempotent by `(userId, idempotencyKey)`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendTextRequest"
            example:
              userId: "1"
              idempotencyKey: "text-123"
              type: normal
              text: "Hello"
              recipients: ["09121234567", "09351234567"]
      responses:
        "200":
          description: Accepted (may include skipped/rejected per-recipient items)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessagesEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /messages:
    get:
      tags: [Messages]
      summary: List messages with filters
      operationId: listMessages
      parameters:
        - name: userId
          in: query
          schema: { type: integer, format: int64 }
        - name: requestId
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema:
            type: string
            enum: [accepted, queued, sending, sent, failed, rejected]
        - name: type
          in: query
          schema:
            type: string
            enum: [otp, text]
        - name: deliveryMode
          in: query
          schema:
            type: string
            enum: [express, normal]
        - name: recipient
          in: query
          schema: { type: string }
        - name: errorCode
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Message list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"

components:
  parameters:
    UserIdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1

  schemas:
    EnvelopeError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        fields:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }

    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/EnvelopeError"

    User:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        balance: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    UserEnvelope:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/User"

    AddBalanceRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          format: int64
          minimum: 1

    OtpRecipient:
      type: object
      required: [mobile]
      properties:
        mobile: { type: string, example: "09121234567" }
        variables:
          type: object
          additionalProperties: { type: string }

    SendOTPRequest:
      type: object
      required: [userId, idempotencyKey, type, template, recipients]
      properties:
        userId: { type: string, example: "1" }
        idempotencyKey: { type: string, example: "otp-123" }
        type:
          type: string
          enum: [express, normal]
        template: { type: string }
        recipients:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OtpRecipient"

    SendTextRequest:
      type: object
      required: [userId, idempotencyKey, type, text, recipients]
      properties:
        userId: { type: string, example: "1" }
        idempotencyKey: { type: string, example: "text-123" }
        type:
          type: string
          enum: [express, normal]
        text: { type: string }
        recipients:
          type: array
          minItems: 1
          items: { type: string }

    MessageItem:
      type: object
      properties:
        id: { type: integer, format: int64 }
        recipient: { type: string }
        type: { type: string, enum: [otp, text] }
        deliveryMode: { type: string, enum: [express, normal] }
        status:
          type: string
          enum: [accepted, queued, sending, sent, failed, rejected]
        cost: { type: integer, format: int64 }
        errorCode: { type: string }
        text: { type: string }

    SendMessagesResult:
      type: object
      properties:
        requestId: { type: integer, format: int64 }
        totalCost: { type: integer, format: int64 }
        acceptedCount: { type: integer }
        rejectedCount: { type: integer }
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MessageItem"

    SendMessagesEnvelope:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/SendMessagesResult"

    MessageDetail:
      type: object
      properties:
        id: { type: integer, format: int64 }
        requestId: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        recipient: { type: string }
        type: { type: string, enum: [otp, text] }
        deliveryMode: { type: string, enum: [express, normal] }
        text: { type: string }
        status:
          type: string
          enum: [accepted, queued, sending, sent, failed, rejected]
        cost: { type: integer, format: int64 }
        errorCode: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    MessageListEnvelope:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MessageDetail"

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: bad_request
              message: invalid id
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: validation_error
              message: validation failed
              fields:
                - field: amount
                  message: must be > 0
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: not_found
              message: resource not found